refactored df output for simplifying
