+++ src/fasttrackpy/tracks.py
@@ -3,6 +3,7 @@ import numpy as np
 from fasttrackpy.processors.smoothers import Smoother
 from fasttrackpy.processors.losses import Loss
 from fasttrackpy.processors.aggs import Agg
+from fasttrackpy.processors.outputs import to_dataframe
 
 from sklearn.experimental import enable_iterative_imputer
 from sklearn.impute import IterativeImputer
@@ -76,6 +77,7 @@ class OneTrack(Track):
         self.n_measured_formants = self._get_measured()
         self.imputed_formants = self._impute_formants()
         self.smoothed_list = self._smooth_formants()
+        self.to_dataframe = to_dataframe
     
     def __repr__(self):
         return f"A formant track object. {self.formants.shape}"
@@ -145,33 +147,9 @@ class OneTrack(Track):
         error = self.agg_fun.aggregate(msqe)
         return error
     
-    def to_dataframe(self):
-        """Return data as a data frame
-
-        Returns:
-           (pl.DataFrame): A data frame
-        """
-        orig_names = [
-            f"F{x}" for x in np.arange(self.n_measured_formants)+1
-        ]
-        smooth_names = [
-            f"F{x}_s" for x in np.arange(self.n_measured_formants)+1
-        ]
-        orig_df = pl.DataFrame(
-            data = self.formants.T,
-            schema=orig_names
-        )
-        orig_df = orig_df.with_columns(
-            time = pl.lit(self.time_domain)
-        )
-
-        smooth_df = pl.DataFrame(
-            data = self.smoothed_formants.T,
-            schema=smooth_names
-        )
-        out_df = pl.concat([orig_df, smooth_df], how = "horizontal")
-
-        return out_df
+    def to_df(self):
+        return to_dataframe(self)
+    
 
 class CandidateTracks(Track):
     """A class for candidate tracks for a single formant
